+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 ### Added
 - Created a setting panel for the audio player
 - Selection of the audio output device. Closes [#4](https://github.com/AresValley/Artemis/issues/4)
+- Added the following (advanced) products to the space weather module [#53](https://github.com/AresValley/Artemis/issues/53):
+  - Sun Synoptic Analysis (SAS)
+  - GOES Solar Ultraviolet Imager (SUVI), all wavelenght (94 Å, 131 Å, 171 Å, 195 Å, 284 Å, 304 Å)
+  - GOES Thematic Map (from SUVI)
+  - Large Angle and Spectrometric Coronagraph (LASCO), both C2/C3 optical imaging of the Sun’s corona
+  - Helioseismic and Magnetic Imager (HMI): Magnetogram, Intensitygram, Dopplergram (TODO)
 
 ### Fixed
 - Fixed margins for downloader window