+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 ### Added
 - Created a setting panel for the audio player
 - Selection of the audio output device. Closes [#4](https://github.com/AresValley/Artemis/issues/4)
-- Added the following (advanced) products to the space weather module [#53](https://github.com/AresValley/Artemis/issues/53):
+- Added the following (advanced) products to the space weather module. Closes [#53](https://github.com/AresValley/Artemis/issues/53):
   - Sun Synoptic Analysis (SAS)
   - GOES Solar Ultraviolet Imager (SUVI), all wavelenght (94 Å, 131 Å, 171 Å, 195 Å, 284 Å, 304 Å)
   - GOES Thematic Map (from SUVI)